Split parameter_spec.js
